Relative motion fixes over wireless, and posture commands are blocking (#134)
* Compute relative motions dynamically and send them to the robot as
global positions

* wait for arm

* Make sure it properly waits for arm to arrive at configuration before
releasing from "posture" commands

* example of simple nav commands

This app provides an example of running simple navigation commands. It is based on the example given at the start of the main Stretch AI documentation in the Quickstart section.

Currently, this code does not run as expected. I am adding it to help us test these basic navigation commands.

* waiting for observations

* updates

* update to catch posegraph missing

* fixed blocking issue & reduced distance

Navigating to the origin now uses a blocking command. Using a non-blocking command resulted in unintuitive behavior.

This commit also reduces the distance of the relative translation command from half a meter to a quarter of a meter to reduce the likelihood of colliding with something.

* Rename navigate_to to move_base_to

* default to blocking in navigate_to

* update navigate to in readme

* update

* link to example + run

* closing things and making it work better

* remove blockng and set robot ip addr script

* updates

* some minor changes for readability

* increase default "not moving" steps

* update

* updates

* add say actions

* updates

* update messages

* add step messages to the server side

* skip duplicate actions

* skipping duplicated incoming actions

* update

* guarantee messages are sent and received

* add move back to origin at end to make it easier

* Updates to zmq config for the robot

* update zmq - add a wrapper with a lock for sending actions at all

* slightly more conservative padding

* threading

* gui setup

---------

Co-authored-by: Charlie Kemp <ck@hello-robot.com>